Move common layers used by elastic-nn based search spaces to nnabla_nas/contrib/common/ofa/layers.py. Merge similar classes and rename class ambiguous class name(Issue #69). Update import path.
