+++ nnabla_nas/contrib/common/ofa/layers.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2020 Sony Corporation. All Rights Reserved.
+# Copyright (c) 2022 Sony Corporation. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,63 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import math
 from collections import OrderedDict
+import nnabla.functions as F
+from nnabla.initializer import ConstantInitializer, UniformInitializer
 
 from .... import module as Mo
-from .ofa_modules.static_op import SEModule
-from .ofa_utils.common_tools import get_same_padding, min_divisible_value
-
-
-CANDIDATES = {
-    'MB3 3x3': {'ks': 3, 'expand_ratio': 3},
-    'MB3 5x5': {'ks': 5, 'expand_ratio': 3},
-    'MB3 7x7': {'ks': 7, 'expand_ratio': 3},
-    'MB4 3x3': {'ks': 3, 'expand_ratio': 4},
-    'MB4 5x5': {'ks': 5, 'expand_ratio': 4},
-    'MB4 7x7': {'ks': 7, 'expand_ratio': 4},
-    'MB6 3x3': {'ks': 3, 'expand_ratio': 6},
-    'MB6 5x5': {'ks': 5, 'expand_ratio': 6},
-    'MB6 7x7': {'ks': 7, 'expand_ratio': 6},
-    'skip_connect': {'ks': None, 'expand_ratio': None},
-}
-
-
-def candidates2subnetlist(candidates):
-    ks_list = []
-    expand_list = []
-    for candidate in candidates:
-        ks = CANDIDATES[candidate]['ks']
-        e = CANDIDATES[candidate]['expand_ratio']
-        if ks not in ks_list:
-            ks_list.append(ks)
-        if e not in expand_list:
-            expand_list.append(e)
-    return ks_list, expand_list
-
-
-def genotype2subnetlist(op_candidates, genotype):
-    op_candidates.append('skip_connect')
-    subnet_list = [op_candidates[i] for i in genotype]
-    ks_list = [CANDIDATES[subnet]['ks'] if subnet != 'skip_connect'
-               else 3 for subnet in subnet_list]
-    expand_ratio_list = [CANDIDATES[subnet]['expand_ratio'] if subnet != 'skip_connect'
-                         else 4 for subnet in subnet_list]
-    depth_list = []
-    d = 0
-    for i, subnet in enumerate(subnet_list):
-        if subnet == 'skip_connect':
-            if d > 1:
-                depth_list.append(d)
-                d = 0
-        elif d == 4:
-            depth_list.append(d)
-            d = 1
-        elif i == len(subnet_list) - 1:
-            depth_list.append(d + 1)
-        else:
-            d += 1
-    assert([d > 1 for d in depth_list])
-    return ks_list, expand_ratio_list, depth_list
+from .utils.common_tools import get_same_padding, min_divisible_value, make_divisible
+from ....module.initializers import he_initializer, torch_initializer
 
 
 def set_layer_from_config(layer_config):
@@ -331,6 +282,49 @@ class LinearLayer(Mo.Sequential):
         return get_extra_repr(self)
 
 
+class SEModule(Mo.Module):
+
+    r"""Squeeze-and-Excitation module, that adaptively recalibrates channel-wise
+        feature responces by explicitlly modelling interdependencies
+        between channels.
+
+    Args:
+        channel (int): The number of input channels.
+        reduction (int, optional): The reduction rate used to determine
+            the number of middle channels.
+
+    References:
+    [1] Hu, Jie, Li Shen, and Gang Sun. "Squeeze-and-excitation networks."
+        Proceedings of the IEEE conference on computer vision and pattern
+        recognition. 2018.
+    """
+
+    REDUCTION = 4
+
+    def __init__(self, channel, reduction=None, name=''):
+        self._name = name
+        self._scope_name = f'<semodule at {hex(id(self))}>'
+
+        self._channel = channel
+        self.reduction = SEModule.REDUCTION if reduction is None else reduction
+
+        num_mid = make_divisible(self._channel // self.reduction)
+
+        self.fc = Mo.Sequential(OrderedDict([
+            ('reduce', Mo.Conv(
+                self._channel, num_mid, (1, 1), pad=(0, 0), stride=(1, 1), with_bias=True)),
+            ('relu', Mo.ReLU()),
+            ('expand', Mo.Conv(
+                num_mid, self._channel, (1, 1), pad=(0, 0), stride=(1, 1), with_bias=True)),
+            ('h_sigmoid', Mo.Hsigmoid())
+        ]))
+
+    def call(self, input):
+        y = F.mean(input, axis=(2, 3), keepdims=True)
+        y = self.fc(y)
+        return input * y
+
+
 def build_activation(act_func, inplace=False):
     if act_func == 'relu':
         return Mo.ReLU(inplace=inplace)
@@ -342,3 +336,48 @@ def build_activation(act_func, inplace=False):
         return None
     else:
         raise ValueError('do not support: %s' % act_func)
+
+
+def init_models(net, model_init='he_fout'):
+    """ Initilizes parameters of Convolution, BatchNormalization, Linear,"""
+    if isinstance(net, list):
+        for sub_net in net:
+            init_models(sub_net, model_init)
+        return
+    for _, m in net.get_modules():
+        if isinstance(m, Mo.Conv):
+            if model_init == 'he_fout':
+                w_init = he_initializer(m._out_channels, m._kernel[0], rng=None)
+                m._W = Mo.Parameter(
+                    m._W.shape, initializer=w_init, scope=m._scope_name)
+            elif model_init == 'he_fin':
+                w_init = he_initializer(m._in_channels, m._kernel[0], rng=None)
+                m._W = Mo.Parameter(
+                    m._W.shape, initializer=w_init, scope=m._scope_name)
+            elif model_init == 'pytorch':
+                w_init = torch_initializer(
+                    m._in_channels, m._kernel, rng=None)
+                m._W = Mo.Parameter(
+                    m._W.shape, initializer=w_init, scope=m._scope_name)
+            else:
+                raise NotImplementedError
+            if m._b is not None:
+                b_init = ConstantInitializer(0)
+                m._b = Mo.Parameter(
+                    m._b.shape, initializer=b_init, scope=m._scope_name)
+        elif isinstance(m, Mo.BatchNormalization):
+            beta_init = ConstantInitializer(0)
+            m._beta = Mo.Parameter(
+                m._beta.shape, initializer=beta_init, scope=m._scope_name)
+            gamma_init = ConstantInitializer(1)
+            m._gamma = Mo.Parameter(
+                m._gamma.shape, initializer=gamma_init, scope=m._scope_name)
+        elif isinstance(m, Mo.Linear):
+            stdv = 1. / math.sqrt(m._W.shape[1])
+            w_init = UniformInitializer((-stdv, stdv))
+            m._W = Mo.Parameter(
+                m._W.shape, initializer=w_init, scope=m._scope_name)
+            if m._b is not None:
+                b_init = ConstantInitializer(0)
+                m._b = Mo.Parameter(
+                    m._b.shape, initializer=b_init, scope=m._scope_name)
